Address /slmos-review feedback on PR #582
Warning + Suggestions 1, 2, 3, 4 from the review:

* `model_mem_init` now returns `AllocError::Oversized` when either
  pool would exceed `MAX_BLOCKS_PER_POOL × BLOCK_SIZE`. Previously
  the pool's `init` did `count.min(MAX_BLOCKS_PER_POOL)` and silently
  truncated, masking a future regression where `MODEL_MEM_WEIGHT_MB`
  is bumped past 1024 without a matching bump to MAX_BLOCKS_PER_POOL.
  Hard-fails at boot now (printed by `rust_model_mem_init`'s C-side
  match arm) instead of giving the caller a smaller pool than asked.

* New `rust_heap_size_bytes()` FFI getter + `test_rust_heap_size_matches_config`
  smoke test pin the boot-time invariant that the per-platform
  `RUST_HEAP_MB` from `<config.h>` is honored end-to-end. Catches a
  silent regression that would otherwise only surface as an opaque
  `handle_alloc_error` panic mid-decode when the SLM forward path's
  KV cache outgrows a too-small heap.

* `kernel/src/main.c` page-count and byte-count multiplies are now
  explicitly `size_t`-promoted so a hypothetical `RUST_HEAP_MB > 4096`
  cannot wrap `unsigned int` before the size_t cast. Defensive — the
  compile-time assert below caps RUST_HEAP_MB at 4 GiB anyway.

* `_Static_assert(RUST_HEAP_MB <= 4096u, ...)` mirrors the existing
  `MODEL_MEM_*_MB <= 1024u` ceiling for symmetry.

* Cross-reference comment in `<config.h>` near the `MODEL_MEM_*_MB`
  asserts pointing at `runtime/src/mm/model_mem.rs::MAX_BLOCKS_PER_POOL`
  so a maintainer raising the C cap finds the Rust dependency.

QEMU `make test` PASSED (incl. new heap-size test). Re-verified on
jetson-nano-1: pools still report 640 blocks / 1280 MB.

+++ kernel/src/main.c
@@ -476,13 +476,17 @@ void kernel_main(void *dtb)
 
     /* Allocate heap for Rust. RUST_HEAP_MB is per-platform in
      * <config.h> — sized by SLM forward-path demand (KV cache +
-     * ForwardScratch); see the comment block there. */
-    const size_t rust_heap_pages = (size_t)RUST_HEAP_MB * 256u; /* 256 pages = 1 MB */
+     * ForwardScratch); see the comment block there. Both multiplies
+     * are size_t-promoted explicitly so a hypothetical >4 GB heap
+     * (RUST_HEAP_MB > 4096) cannot wrap unsigned int before the
+     * promotion. The compile-time assert in <config.h> caps
+     * RUST_HEAP_MB well below that, but the cast is cheap insurance. */
+    const size_t rust_heap_pages = (size_t)RUST_HEAP_MB * (size_t)256; /* 256 pages = 1 MB */
     void *rust_heap = pmm_alloc_pages(rust_heap_pages);
     if (!rust_heap) {
         panic("Failed to allocate Rust heap (%u MB requested)", RUST_HEAP_MB);
     }
-    rust_heap_init(rust_heap, rust_heap_pages * 4096u);
+    rust_heap_init(rust_heap, rust_heap_pages * (size_t)4096);
     INFO("  Rust heap: %p (%u MB)", rust_heap, RUST_HEAP_MB);
 
     /* Call Rust init and verify */
